# ffs-api/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API for user authentication, creation, and management of schools, districts, bids, and tasks"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get list of all users",
        "description": "Retrieve a list of all users in the system. Accessible only by Super-Admin, Group Admin, and District Admin roles.",
        "tags": ["User"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListUsersResponseItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListUsersResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleAssignment"
            },
            "description": "User's platform role assignments"
          },
          "bidRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BidRoleAssignment"
            },
            "description": "User's bid-specific role assignments"
          },
          "managedBids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of bid IDs that the user manages"
          },
          "status": {
            "type": "string",
            "description": "User's current status"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "User's last login timestamp in ISO format"
          },
          "demoAccount": {
            "type": "boolean",
            "description": "Whether this is a demo account"
          }
        },
        "required": ["id", "firstName", "lastName", "email", "roles", "bidRoles", "managedBids", "status", "lastLogin", "demoAccount"]
      },
      "RoleAssignment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Role type name"
          },
          "scope": {
            "$ref": "#/components/schemas/RoleScope"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of permission names"
          }
        },
        "required": ["type", "scope", "permissions"]
      },
      "BidRoleAssignment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Bid role type name"
          },
          "scope": {
            "$ref": "#/components/schemas/RoleScope"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of permission names"
          }
        },
        "required": ["type", "scope", "permissions"]
      },
      "RoleScope": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Scope type (e.g., 'coop', 'district', 'school')"
          },
          "id": {
            "type": "string",
            "description": "Scope identifier"
          },
          "name": {
            "type": "string",
            "description": "Optional scope name for display"
          }
        },
        "required": ["type", "id"]
      },
...
    }
  }
}